requirements.txt
earthengine-api>=0.1.347
google-api-python-client>=2.70.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
rasterio>=1.3.6
fiona>=1.9.4
shapely>=2.0.1
numpy>=1.25.0
scipy>=1.11.3
scikit-image>=0.21.0
matplotlib>=3.8.1
torch>=2.2.0
torchvision>=0.17.0
mlflow>=2.6.1
dvc>=2.40.0
fastapi>=0.98.0
uvicorn>=0.22.0
google-auth-httplib2>=0.1.0
google-auth-httplib2>=0.1.0
google-auth-httplib2>=0.1.0
tqdm>=4.66.1
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.70.0
